OpenAL
Model, View combine

+++ AppTemplate/inc/context.h
@@ -20,8 +20,8 @@ public:
 	Context();
 	~Context();
 	void ResetSource();
-	void setSourcePos(ALuint source, float x, float y, float z);
-	void setListenerPos(float x, float y, float z);
+	void setSourcePos(ALuint source, int x, int y, int z);
+	void setListenerPos(int x, int y, int z);
 	void Play();
 	void Stop();
 	void Push(ALuint source);
